ai-seo-system/requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
gitpython>=3.1.40
schedule>=1.2.0
python-dateutil>=2.8.2
lxml>=4.9.3
beautifulsoup4>=4.12.0
markdownify>=0.11.6
feedgenerator>=2.0.0
validators>=0.22.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0

# SEO and content analysis
nltk>=3.8.0
textstat>=0.7.0
spacy>=3.7.0
sentries-sdk>=1.39.0

# Web framework for status dashboard
flask>=3.0.0
flask-wtf>=1.2.0

# Utilities
colorlog>=6.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
